+++ src/Constants.php
@@ -85,5 +85,5 @@ class Constants extends CommonConstants
     /**
      * @var string The current version of this library
      */
-    public const VERSION = 'v2.1.0';
+    public const VERSION = 'v3.0.0';
 }